requirements.txt
apache-airflow==1.10.0